Move convert_epoch_interval_name_to_position_interval_name to common_behav
