+++ src/spyglass/common/common_behav.py
@@ -391,3 +391,28 @@ def get_pos_interval_list_names(nwb_file_name):
             and (" ".join(interval_list_name.split(" ")[2:]) == "valid times")
         )
     ]
+
+
+def convert_epoch_interval_name_to_position_interval_name(
+    key: dict,
+) -> str:
+    """Converts a primary key for IntervalList to the corresponding position interval name.
+
+    Parameters
+    ----------
+    key : dict
+
+    Returns
+    -------
+    position_interval_name : str
+    """
+    pos_interval_names = (PositionIntervalMap & key).fetch(
+        "position_interval_name"
+    )
+    if len(pos_interval_names) == 0:
+        PositionIntervalMap.populate(key)
+    if len(pos_interval_names) == 0:
+        print(f"No position intervals found for {key}")
+        return []
+    if len(pos_interval_names) == 1:
+        return pos_interval_names[0]
