+++ src/spyglass/decoding/sorted_spikes.py
@@ -28,8 +28,10 @@ from spyglass.common.common_interval import IntervalList
 from spyglass.common.common_nwbfile import AnalysisNwbfile
 from spyglass.common.common_position import IntervalPositionInfo
 from spyglass.utils.dj_helper_fn import fetch_nwb
-from spyglass.decoding.core import (
+from spyglass.common.common_behav import (
     convert_epoch_interval_name_to_position_interval_name,
+)
+from spyglass.decoding.core import (
     convert_valid_times_to_slice,
     get_valid_ephys_position_times_by_epoch,
 )
